Fixed the bug
Fixed the bug
After the game was over, if you still typed something on the usertextbox, the timer would start again.

Added a boolean named check and set it true. Added an if statement to check whether "check" is true or not. If it is true start the timer. When the game is over i set "check" false so that even if you type timer wont start. When you click on restart button i set check to true again and now the timer starts when you type :D